Don't strip newlines from the end of tags since this clobbers pre-existing formatting.
This is particularly egregious if you are running something through textile and want
to be able to strip tags to get nicely formatted plain text.  As far as I can tell
that code was just a workaround for a nokogiri (or possibly libxml2) bug that I isolated
in a test case at http://github.com/tenderlove/nokogiri/issues#issue/149

Rather than encode the broken behavior in tests I simply changed the markup slightly to
work around the issue.